Tests: Move promise/A+ adapters for tests to dedicated folder
